pyproject.toml
[project]
name = "baseball-hydra"
version = "0.1.0"
description = "MLB batter next-season stat predictions using multi-task neural networks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "pybaseball>=2.2.7",
    "pyyaml>=6.0",
    "scikit-learn>=1.6",
    "torch>=2.2",
    "matplotlib>=3.8",
    "seaborn>=0.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
